# binno_backend/contracts/binno-openapi-v1.yaml
openapi: 3.1.0
info:
  title: BINNO Platform API
  version: 1.1.0
  description: |
    BINNO MVP contract-first API. All IDs are UUID v7, all timestamps are
    UTC, all money is integer tiyin. Mutations require an Idempotency-Key.
servers:
  - url: https://api.binno.uz/api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/otp/request:
    post:
      summary: Send a one-time code by SMS
      description: >-
        Answers the same way whether or not the number has an account.
        Distinguishing them would turn this into a membership oracle for every
        phone number in the country. Registration happens on first successful
        verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: >-
                    Uzbek number in any common form. Normalised to 998XXXXXXXXX
                    before storage, so +998 90 123 45 67 and 901234567 are the
                    same account.
                  example: "+998901234567"
      responses:
        '202':
          description: >-
            Handed to the SMS gateway. Not 200, whether it reaches the handset
            is not something this response can honestly claim.
        '422':
          description: >-
            The number is not a valid Uzbek number, or the SMS gateway will
            never deliver to it. Both are the caller's to fix and neither
            improves on a retry, so they answer alike; a gateway that is merely
            unavailable is a 5xx instead.
        '429':
          description: A code was sent recently; Retry-After gives the wait.
          headers:
            Retry-After: { schema: { type: integer } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/otp/verify:
    post:
      summary: Exchange a code for a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string, example: "+998901234567" }
                code:  { type: string, example: "482913" }
      responses:
        '200':
          description: >-
            Session issued. First successful verification creates the account,
            and is_new_user says whether this call was that one.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      summary: Rotate a session
      description: >-
        Every refresh issues a new pair and revokes the presented token. A token
        replayed after rotation is treated as theft and ends every session for
        that user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New pair issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      summary: End one session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204':
          description: >-
            The session will not work. An unknown, expired or already-revoked
            token also returns 204: the caller's intent is satisfied either way,
            and reporting the difference would let an attacker probe for valid
            tokens.
        default: { $ref: '#/components/responses/Problem' }
...
  /catalog/requests:
    post:
      tags: [catalog]
      summary: Ask for a product to be added to the catalogue
      operationId: createCatalogRequest
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogRequestInput' }
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedResource' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
  /catalog/requests/{id}:
    patch:
      tags: [operator]
      summary: Close a catalogue request as added or rejected
      operationId: resolveCatalogRequest
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogResolution' }
      responses:
        '200': { description: Request closed }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
  /offers:
    post:
      tags: [offers]
      summary: Create a store offer for a master-catalogue product
      operationId: createOffer
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfferInput' }
      responses:
        '201':
          description: Offer created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedResource' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
  /offers/{id}:
    patch:
      tags: [offers]
      summary: Update price, stock, or visibility (pause/republish)
      operationId: updateOffer
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfferUpdate' }
      responses:
        '200': { description: Offer updated }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
...
  /orders/{id}/payment-invoice:
    get:
      tags: [billing]
      summary: The immutable payment invoice PDF
      operationId: getPaymentInvoice
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
  /orders/{id}/refund:
    post:
      tags: [billing]
      summary: The seller submits refund evidence
      operationId: submitRefundEvidence
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefundEvidence' }
      responses:
        '200': { description: Evidence submitted }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
...
  /owners/{id}/commission-invoices:
    get:
      tags: [billing]
      summary: Owner-level weekly commission invoices
      operationId: listCommissionInvoices
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Weekly invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CommissionInvoice' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. Required claims:
          * `sub`: the user's UUID
          * `exp`: mandatory; a token without exp is rejected outright
          * `roles`: array of `buyer` | `seller` | `operator`;
            unknown roles are dropped

        Authentication (is the token genuine) and authorisation (is this
        resource yours) are separate concerns. A genuine token is not enough:
        every endpoint checks the resource's owner on its own. Requesting
        another user's order or owner account answers 403, or 404 where
        existence must not be disclosed.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Server-issued opaque composite keyset cursor; the client sends it back without parsing it
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
      description: |
        Retry-safe mutation key with a 24-hour window.

        Same key + same body -> the first response is replayed verbatim
        (`Idempotency-Replayed: true`).
        Same key + a DIFFERENT body -> 422, `type` =
        `https://binno.uz/problems/idempotency-key-reuse`. The mutation does
        not run. Use a fresh key for every logically distinct request.
...
  responses:
    Problem:
      description: RFC 7807 error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Illegal state transition or duplicate action
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Validation:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Token missing, malformed, or expired
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: |
        The token is genuine but the resource belongs to another account.
        Where disclosing existence is a risk (for example someone else's
        order), 404 is returned instead.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: |
        The caller's request budget is exhausted. Authenticated callers are
        counted per principal, anonymous callers per IP, and the anonymous
        budget is smaller.
      headers:
        Retry-After:
          description: Seconds until the next window
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    Money:
      type: integer
      format: int64
      minimum: 0
      description: UZS tiyin
    Quantity:
      type: string
      pattern: '^[0-9]+(?:\.[0-9]+)?$'
      description: Decimal quantity carried as a JSON string
    CreatedResource:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    CatalogRequestInput:
      type: object
      required: [store_id, name, unit]
      properties:
        store_id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 200 }
        unit: { type: string, minLength: 1, maxLength: 30 }
        description: { type: string, maxLength: 2000 }
        image_url: { type: string, format: uri }
    CatalogResolution:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [added, rejected] }
        product_id: { type: string, format: uuid }
        reject_reason: { type: string, maxLength: 1000 }
    OfferInput:
      type: object
      required: [store_id, product_id, price, declared_qty]
      properties:
        store_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        price: { $ref: '#/components/schemas/Money' }
        declared_qty: { $ref: '#/components/schemas/Quantity' }
    OfferUpdate:
      type: object
      minProperties: 1
      properties:
        price: { $ref: '#/components/schemas/Money' }
        declared_qty: { $ref: '#/components/schemas/Quantity' }
        status:
          type: string
          enum: [published, hidden]
          description: >-
            Pause the offer (hidden) or republish it (published). An archived
            offer cannot be brought back from here.
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: >-
            Short-lived HS256 JWT carrying sub and roles. Stateless, so nothing
            can revoke it before it expires, which is why it is short.
        refresh_token:
          type: string
          description: >-
            Opaque, long-lived, single-use. Stored hashed, so it can be revoked
            and so a database dump yields no usable session.
        token_type: { type: string, enum: [Bearer] }
        expires_in:
          type: integer
          description: Access token lifetime in seconds.
    Session:
      description: >-
        A verified session: the token pair plus the one thing only verification
        can report. Refresh returns a bare TokenPair, because a rotation has no
        opinion on how the account began.
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          required: [is_new_user]
          properties:
            is_new_user:
              type: boolean
              description: >-
                True when this verification created the account, so the client
                can route a first-time user into onboarding rather than to the
                feed. Answered here and not at /auth/otp/request on purpose: the
                same fact, told before a code is verified, is a membership
                oracle for every phone number in the country. Told to someone
                who has just proved possession of the handset, it discloses
                nothing they do not already know.
...
    RefundEvidence:
      type: object
      required: [amount, evidence_url]
      properties:
        amount: { $ref: '#/components/schemas/Money' }
        evidence_url: { type: string, format: uri }
...
    CommissionInvoice:
      type: object
      required:
        - id
        - period_start
        - period_end
        - accrued_amount
        - discount_amount
        - amount
        - due_at
        - status
      properties:
        id: { type: string, format: uuid }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        accrued_amount: { $ref: '#/components/schemas/Money' }
        discount_amount: { $ref: '#/components/schemas/Money' }
        amount: { $ref: '#/components/schemas/Money' }
        due_at: { type: string, format: date-time }
        status: { type: string, enum: [issued, paid, overdue] }
...